Fix game rules and improve performance

+++ src/modules/agent-deep-q-scratch-built/lib-deep-q-network-scratch-built/Agent.js
@@ -25,7 +25,7 @@ export default class Agent {
         this._lastActionStats = {
             action: 0,
             wasRandom: false,
-            weights: new Float32Array(numberOfActions),
+            weights: new Float64Array(numberOfActions),
             tdError: 0
         };
